non-generic ImageFrameCollection API definition
