+++ src/ImageSharp/Image{TPixel}.cs
@@ -149,7 +149,7 @@ namespace SixLabors.ImageSharp
         /// <returns>Returns a new <see cref="Image{TPixel}"/> with all the same pixel data as the original.</returns>
         public Image<TPixel> Clone(Configuration configuration)
         {
-            IEnumerable<ImageFrame<TPixel>> clonedFrames = this.Frames.Select(x => x.Clone(configuration));
+            IEnumerable<ImageFrame<TPixel>> clonedFrames = this.Frames.Select<ImageFrame<TPixel>,ImageFrame<TPixel>>(x => x.Clone(configuration));
             return new Image<TPixel>(configuration, this.Metadata.DeepClone(), clonedFrames);
         }
 
@@ -161,7 +161,7 @@ namespace SixLabors.ImageSharp
         /// <returns>The <see cref="Image{TPixel2}"/>.</returns>
         public override Image<TPixel2> CloneAs<TPixel2>(Configuration configuration)
         {
-            IEnumerable<ImageFrame<TPixel2>> clonedFrames = this.Frames.Select(x => x.CloneAs<TPixel2>(configuration));
+            IEnumerable<ImageFrame<TPixel2>> clonedFrames = this.Frames.Select<ImageFrame<TPixel>,ImageFrame<TPixel2>>(x => x.CloneAs<TPixel2>(configuration));
             return new Image<TPixel2>(configuration, this.Metadata.DeepClone(), clonedFrames);
         }
